refactor: split CommandCatalog into a DOM-free command_catalog module
Extract the catalog — CommandCatalog, CommandCatalogContext,
collectActionBindings, the CommandPaletteEntry types, and the tool/label
helpers — from command_palette.ts into a new command_catalog.ts with no DOM
or CSS dependencies. command_palette.ts keeps the Overlay-based CommandPalette
UI and bindCommandPalette, importing the catalog and the stylesheet.

Previously, importing CommandCatalog for its enumeration transitively pulled
in command_palette.css and the Overlay class even when no palette was
rendered. Splitting the modules lets the catalog be consumed (and unit-tested)
without a DOM, and reused independently of the palette UI.

- command_catalog.spec.ts (renamed from command_palette.spec.ts) now imports
  from command_catalog.js, so the catalog tests no longer depend on the
  palette module.
- default_viewer_setup.ts imports CommandCatalog from command_catalog.js and
  bindCommandPalette from command_palette.js.

No behavioural change.
